+++ src/util.c
@@ -1,5 +1,6 @@
 #include <ctype.h>
 #include <limits.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <string.h>
 
@@ -11,7 +12,7 @@ char *strtolower(const char *s) {
   return rv;
 }
 
-int find_char_idx(const char *s, char c) {
+int16_t find_char_idx(const char *s, char c) {
   for (size_t i = 0; i < strlen(s); i++) {
     if (s[i] == c) {
       return i;
@@ -20,9 +21,9 @@ int find_char_idx(const char *s, char c) {
   return -1;
 }
 
-unsigned short get_num_strlen(size_t num) {
-  char buf[USHRT_MAX];
-  memset(buf, 0, USHRT_MAX);
+uint8_t get_num_strlen(size_t num) {
+  char buf[UINT8_MAX];
+  memset(buf, 0, UINT8_MAX);
   sprintf(buf, "%zd", num);
   return strlen(buf);
 }
